+++ app/helpers/responsive-image-url.js
@@ -1,1 +1,0 @@
-export { default, responsiveImageUrl } from 'empress-blog-casper-template/helpers/responsive-image-url';